#71: Change generated by green dao files to fit coday, Change task template activity to landscape always, renames
