--- Friendly-plans/app/src/androidTest/java/pg/autyzm/friendly_plans/ApplicationTest.java
@@ -1,10 +1,0 @@
-package pg.autyzm.friendly_plans;
-
-import android.app.Application;
-import android.test.ApplicationTestCase;
-
-public class ApplicationTest extends ApplicationTestCase<Application> {
-    public ApplicationTest() {
-        super(Application.class);
-    }
-}
