+++ Friendly-plans/app/src/main/AndroidManifest.xml
@@ -4,9 +4,9 @@
 
   <application
     android:allowBackup="true"
-    android:name=".App"
     android:icon="@mipmap/ic_launcher"
     android:label="@string/app_name"
+    android:name=".App"
     android:supportsRtl="true"
     android:theme="@style/AppTheme">
     <activity
@@ -14,11 +14,13 @@
       android:screenOrientation="landscape">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity android:name=".TaskCreateActivity">
-        </activity>
-    </application>
+        <category android:name="android.intent.category.LAUNCHER"/>
+      </intent-filter>
+    </activity>
+    <activity
+      android:name=".TaskCreateActivity"
+      android:screenOrientation="landscape">
+    </activity>
+  </application>
 
 </manifest>
