+++ Friendly-plans/app/src/main/java/database/entities/TaskTemplate.java
@@ -1,26 +1,31 @@
-package database;
+package database.entities;
 
 import org.greenrobot.greendao.annotation.Entity;
 import org.greenrobot.greendao.annotation.Id;
 import org.greenrobot.greendao.annotation.ToMany;
 
 import java.util.List;
-
 import org.greenrobot.greendao.annotation.Generated;
 import org.greenrobot.greendao.DaoException;
 
 @Entity
-public class PlanTemplate {
+public class TaskTemplate {
 
     @Id(autoincrement = true)
     private Long id;
 
     private String name;
 
-    @ToMany(referencedJoinProperty = "childId")
-    private List<ChildPlan> childPlans;
+    private String picture;
 
-    @ToMany(referencedJoinProperty = "planTemplateId")
+    private String sound;
+
+    private int durationTime;
+
+    @ToMany(referencedJoinProperty = "taskTemplateId")
+    private List<StepTemplate> stepTemplates;
+
+    @ToMany(referencedJoinProperty = "taskTemplateId")
     private List<PlanTaskTemplate> planTaskTemplates;
 
     /**
@@ -32,19 +37,21 @@ public class PlanTemplate {
     /**
      * Used for active entity operations.
      */
-    @Generated(hash = 667849288)
-    private transient PlanTemplateDao myDao;
+    @Generated(hash = 404230972)
+    private transient TaskTemplateDao myDao;
 
-    @Generated(hash = 643907227)
-    public PlanTemplate(Long id, String name) {
+    @Generated(hash = 1930893288)
+    public TaskTemplate(Long id, String name, String picture, String sound, int durationTime) {
         this.id = id;
         this.name = name;
+        this.picture = picture;
+        this.sound = sound;
+        this.durationTime = durationTime;
     }
 
-    @Generated(hash = 623023216)
-    public PlanTemplate() {
+    @Generated(hash = 2000532247)
+    public TaskTemplate() {
     }
-
 
     public Long getId() {
         return this.id;
@@ -62,42 +69,66 @@ public class PlanTemplate {
         this.name = name;
     }
 
+    public String getPicture() {
+        return this.picture;
+    }
+
+    public void setPicture(String picture) {
+        this.picture = picture;
+    }
+
+    public String getSound() {
+        return this.sound;
+    }
+
+    public void setSound(String sound) {
+        this.sound = sound;
+    }
+
+    public int getDurationTime() {
+        return this.durationTime;
+    }
+
+    public void setDurationTime(int durationTime) {
+        this.durationTime = durationTime;
+    }
+
     /**
      * To-many relationship, resolved on first access (and after reset). Changes to to-many
      * relations are not persisted, make changes to the target entity.
      */
-    @Generated(hash = 339806248)
-    public List<ChildPlan> getChildPlans() {
-        if (childPlans == null) {
+    @Generated(hash = 856664114)
+    public List<StepTemplate> getStepTemplates() {
+        if (stepTemplates == null) {
             final DaoSession daoSession = this.daoSession;
             if (daoSession == null) {
                 throw new DaoException("Entity is detached from DAO context");
             }
-            ChildPlanDao targetDao = daoSession.getChildPlanDao();
-            List<ChildPlan> childPlansNew = targetDao
-                    ._queryPlanTemplate_ChildPlans(id);
+            StepTemplateDao targetDao = daoSession.getStepTemplateDao();
+            List<StepTemplate> stepTemplatesNew = targetDao
+                    ._queryTaskTemplate_StepTemplates(id);
             synchronized (this) {
-                if (childPlans == null) {
-                    childPlans = childPlansNew;
+                if (stepTemplates == null) {
+                    stepTemplates = stepTemplatesNew;
                 }
             }
         }
-        return childPlans;
+        return stepTemplates;
     }
 
     /**
      * Resets a to-many relationship, making the next get call to query for a fresh result.
      */
-    @Generated(hash = 398907355)
-    public synchronized void resetChildPlans() {
-        childPlans = null;
+    @Generated(hash = 664303899)
+    public synchronized void resetStepTemplates() {
+        stepTemplates = null;
     }
 
     /**
      * To-many relationship, resolved on first access (and after reset). Changes to to-many
      * relations are not persisted, make changes to the target entity.
      */
-    @Generated(hash = 710197037)
+    @Generated(hash = 613273513)
     public List<PlanTaskTemplate> getPlanTaskTemplates() {
         if (planTaskTemplates == null) {
             final DaoSession daoSession = this.daoSession;
@@ -106,7 +137,7 @@ public class PlanTemplate {
             }
             PlanTaskTemplateDao targetDao = daoSession.getPlanTaskTemplateDao();
             List<PlanTaskTemplate> planTaskTemplatesNew = targetDao
-                    ._queryPlanTemplate_PlanTaskTemplates(id);
+                    ._queryTaskTemplate_PlanTaskTemplates(id);
             synchronized (this) {
                 if (planTaskTemplates == null) {
                     planTaskTemplates = planTaskTemplatesNew;
@@ -161,10 +192,9 @@ public class PlanTemplate {
     }
 
     /** called by internal mechanisms, do not call yourself. */
-    @Generated(hash = 1753738323)
+    @Generated(hash = 1588370016)
     public void __setDaoSession(DaoSession daoSession) {
         this.daoSession = daoSession;
-        myDao = daoSession != null ? daoSession.getPlanTemplateDao() : null;
+        myDao = daoSession != null ? daoSession.getTaskTemplateDao() : null;
     }
-
 }
