+++ Friendly-plans/app/src/main/java/database/repository/RepositoryComponent.java
@@ -1,4 +1,4 @@
-package dao;
+package database.repository;
 
 import dagger.Component;
 import javax.inject.Singleton;
@@ -6,7 +6,7 @@ import pg.autyzm.friendly_plans.TaskContainerFragment;
 
 @Singleton
 @Component(modules = {DaoSessionModule.class, RepositoryModule.class})
-public interface DaoSessionComponent {
+public interface RepositoryComponent {
 
     @SuppressWarnings("unused")
     TaskTemplateRepository taskTemplateRepository();
